Various compiler warnings removed. Most are of the form "variable x may be used uninitialised" and most of the time it is untrue, however there are a few legitimate coding errors fixed too.
git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3740 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap5/export_snps.c
@@ -146,7 +146,7 @@ static int export_snps_contig(GapIO *io, tg_rec contig, int start, int end,
 		    ref_entry *e;
 		    //e = ref->ref_id[ref_id[i]];
 		    e = ref->ref_id[22]; // HACK FOR TESTING
-		    printf("Loading %d..%d\n", start_pos, e->length);
+		    printf("Loading %d..%d\n", start_pos, (int)e->length);
 		    ref_seq = load_ref_portion(ref->fp, e, start_pos, e->length);
 		}
